requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.4.2
colorama==0.4.4
decorator==5.0.5
distlib==0.3.1
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.4
Pillow==8.2.0
pylint==2.6.0
python-http-client==3.3.2
pytz==2021.1
self==2020.12.3
sendgrid==6.6.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
toml==0.10.2
virtualenv==20.2.1
virtualenvwrapper-win==1.2.6
wrapt==1.12.1
